Refactor base fragments to `org.ole.planet.myplanet.ui.base` package
Moved `BaseDashboardFragment`, `BaseTeamFragment`, and `BaseVoicesFragment` to the `ui/base` package to improve code organization.
Updated package declarations and imports in all affected files.
Removed circular dependency between `BaseTeamFragment` and `BaseVoicesFragment` (as they are now in the same package).
Added necessary imports for classes that were previously in the same package (e.g., `DashboardViewModel`, `BaseDashboardFragmentPlugin`).
