+++ app/src/main/java/org/ole/planet/myplanet/ui/base/BaseTeamFragment.kt
@@ -1,4 +1,4 @@
-package org.ole.planet.myplanet.ui.teams
+package org.ole.planet.myplanet.ui.base
 
 import android.os.Bundle
 import androidx.lifecycle.lifecycleScope
@@ -14,7 +14,6 @@ import org.ole.planet.myplanet.model.RealmMyTeam
 import org.ole.planet.myplanet.model.RealmNews
 import org.ole.planet.myplanet.model.RealmUserModel
 import org.ole.planet.myplanet.repository.TeamsRepository
-import org.ole.planet.myplanet.ui.voices.BaseVoicesFragment
 
 @AndroidEntryPoint
 abstract class BaseTeamFragment : BaseVoicesFragment() {
